Implement basic resource manager
For now, the resource manager is doing some unnecessary work. The goal
was to make something that'd be as easy as possible to adapt in the
future for things such as sprite packing or animated (and other kinds of
sprites).

The window is also no longer responsible for bliting sprites onto the
screen surface, but rather sprites know how to draw themselves onto
arbitrary surfaces (such as the screen).