Use hardware rendering
Considering creating a seperate renderer class. I'm not entirely
convinced that Window should own the renderer as I'm slightly
uncomfortable providing getters to private data simply so the containing
class can pass it around to other componenets. If instead client were to
own the renderer then it'd simply be passing its own data to its
subcomponents.

If there's no reason for Window to own the renderer, I may move it to
the client.